# openapi/pegasus-v1.json
{
  "components": {
    "schemas": {
      "AcceptVehicleSuggestionRequest": {
        "properties": {
          "correction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VehicleCorrectionRequest"
              }
            ]
          },
          "decision": {
            "enum": [
              "accept",
              "correct"
            ],
            "type": "string"
          },
          "editLeaseToken": {
            "type": "string"
          },
          "expectedVersion": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "operationKey": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "expectedVersion",
          "decision",
          "operationKey",
          "reason",
          "editLeaseToken"
        ],
        "type": "object"
      },
      "AcceptedVehicleSuggestionResponse": {
        "properties": {
          "caseId": {
            "format": "uuid",
            "type": "string"
          },
          "confirmationId": {
            "format": "uuid",
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "decision": {
            "enum": [
              "accept",
              "correct"
            ],
            "type": "string"
          },
          "lookupObservationId": {
            "format": "uuid",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/VehicleEvidenceProvenanceResponse"
          },
          "resultingCaseVersion": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "values": {
            "$ref": "#/components/schemas/VehicleConfirmationValuesResponse"
          }
        },
        "required": [
          "confirmationId",
          "caseId",
          "lookupObservationId",
          "decision",
          "values",
          "provenance",
          "resultingCaseVersion",
          "correlationId"
        ],
        "type": "object"
      },
      "CaseVehicleResponse": {
        "properties": {
          "caseId": {
            "format": "uuid",
            "type": "string"
          },
          "confirmationHistory": {
            "items": {
              "$ref": "#/components/schemas/VehicleConfirmationHistoryResponse"
            },
            "type": "array"
          },
          "confirmed": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConfirmedVehicleEvidenceResponse"
              }
            ]
          },
          "correlationId": {
            "type": "string"
          },
          "latestObservation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VehicleLookupObservationResponse"
              }
            ]
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/VehicleLookupObservationResponse"
            },
            "type": "array"
          },
          "version": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "caseId",
          "version",
          "confirmed",
          "latestObservation",
          "observations",
          "confirmationHistory",
          "correlationId"
        ],
        "type": "object"
      },
...
      "PegasusProblem": {},
...
      "VehicleLookupRequest": {
        "properties": {
          "editLeaseToken": {
            "type": "string"
          },
          "expectedVersion": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "operationKey": {
            "type": "string"
          },
          "registration": {
            "type": "string"
          }
        },
        "required": [
          "registration",
          "expectedVersion",
          "operationKey",
          "editLeaseToken"
        ],
        "type": "object"
      },
      "VehicleLookupResponse": {
        "properties": {
          "caseId": {
            "format": "uuid",
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "registration": {
            "type": "string"
          },
          "resultingCaseVersion": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "state": {
            "enum": [
              "pending",
              "processing",
              "retryScheduled",
              "completed",
              "failed",
              "poisoned"
            ],
            "type": "string"
          },
          "workItemId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "workItemId",
          "caseId",
          "registration",
          "state",
          "resultingCaseVersion",
          "correlationId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Versioned JSON API for the Pegasus native desktop (product version 0.1.0-alpha.1).",
    "title": "Pegasus Desktop Gateway",
    "version": "0.1.0-alpha.1"
  },
  "openapi": "3.1.1",
  "paths": {
    "/api/v1/cases/{caseId}/vehicle": {
      "get": {
        "description": "Returns confirmed vehicle evidence, lookup observations and provenance.",
        "operationId": "GetVehicleEvidence",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The weak evidence version previously returned by this endpoint.",
            "in": "header",
            "name": "If-None-Match",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseVehicleResponse"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Weak version of the vehicle evidence projection.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified",
            "headers": {
              "ETag": {
                "description": "Weak version of the vehicle evidence projection.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Get vehicle evidence",
        "tags": [
          "VehicleEndpoints"
        ]
      }
    },
    "/api/v1/cases/{caseId}/vehicle/lookups": {
      "post": {
        "description": "Queues a provider lookup through the server-side vehicle workflow.",
        "operationId": "RequestVehicleLookup",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleLookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleLookupResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Request a vehicle lookup",
        "tags": [
          "VehicleEndpoints"
        ]
      }
    },
    "/api/v1/cases/{caseId}/vehicle/suggestions/{suggestionId}/accept": {
      "post": {
        "description": "Accepts or corrects a provider vehicle suggestion through Core.",
        "operationId": "AcceptVehicleSuggestion",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "suggestionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptVehicleSuggestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedVehicleSuggestionResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/PegasusProblem"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Accept a vehicle suggestion",
        "tags": [
          "VehicleEndpoints"
        ]
      }
    }
  },
...
}